week-5/javascript module/python blue print
