+++ taipei-day-trip/api/connector.py
@@ -1,0 +1,20 @@
+
+
+# attraction_api = Blueprint("attraction_api",__name__)
+
+
+# #insert connector.pooling 
+# dbconfig = {
+#     "user" : "root",
+#     "password" : "tian0426",
+#     "host" : "localhost",
+#     "database" : "taipeiDayTrip",
+# }
+
+# connection_pool = mysql.connector.pooling.MySQLConnectionPool(
+#     pool_name = "wehelp_pool",
+#     pool_size = 5,
+#     pool_reset_session = True,
+#     **dbconfig
+# )
+
