+++ taipei-day-trip/static/all.js
@@ -1,3 +1,12 @@
+import navbar_signin_lib from "./navbar_signin_lib.js" 
+navbar_signin_lib()
+
+// import cookie_lib from "./cookie_lib.js"
+// cookie_lib()
+
+import navbar_booking_lib from "./navbar_booking_lib.js"
+navbar_booking_lib()
+
 let nextPage = 0; //預設page = 0
 let urlPage = "";
 let urlDemo = "/api/attractions?page=";
@@ -22,16 +31,16 @@ if (nextPage == 0) {
   getData();
 }
 
+//get datas2
 function getData() {
   fetch(urlPage)
     .then(function (response) {
       return response.json();
-    })
-    .then(function (data) {
+    }).then(function (data) {
       if (data["error"] == true) {
         console.log("error page");
-        return "error page";
-      } else {
+        return "error page";}
+      else {
         attractions = data["data"];
         attractionLen = attractions.length;
         loadPicture();
@@ -59,7 +68,7 @@ function getData() {
 
 //loading picture
 function loadPicture() {
-  for (let i = 0; i < attractionLen; i++) {
+  for(let i = 0; i < attractionLen; i++) {
     let item = document.createElement("div");
     item.setAttribute("class", "item");
     //a 連結
@@ -75,13 +84,12 @@ function loadPicture() {
 
     let imgBoxes = document.querySelector(".imgBoxes");
     imgBoxes.appendChild(itemId);
-    // let imgBoxes = document.querySelector('.imgBoxes')
-    // imgBoxes.appendChild(item)
 
     attractionImg = attractions[i]["images"][0];
     attractionName = attractions[i]["name"];
     attractionMrt = attractions[i]["mrt"];
     attractionCategory = attractions[i]["category"];
+
     // id
     attractionId = attractions[i]["id"];
     attractionIdUrl = "/attraction/" + attractionId;
@@ -113,41 +121,38 @@ function loadPicture() {
 function loadMore() {
   if (judge == 1) {
     const loading = document.querySelector(".loading");
-    window.addEventListener("scroll", () => {
-      const scrollable =
-        document.documentElement.scrollHeight - window.innerHeight;
-      const scrolled = window.scrollY;
+      window.addEventListener("scroll", () => {
+    const scrollable = document.documentElement.scrollHeight - window.innerHeight;
+    const scrolled = window.scrollY;
+    //提早感應
+    if (Math.ceil(scrolled) > scrollable - 50 && judge == 1) {
+      judge = 0;
+      showLoading();
+    }
+  });
 
-      //提早感應
-      if (Math.ceil(scrolled) > scrollable - 50 && judge == 1) {
-        judge = 0;
-        showLoading();
+  function showLoading() {
+    setTimeout(getPost(), 2000);
+  }
+
+  function getPost() {
+    fetch(urlPage)
+      .then(function (response) {
+        return response.json();
+    }).then(function (data) {
+      attractions = data["data"];
+      attractionLen = data["data"].length;
+      loadPicture();
+
+      if (data["nextPage"] != null) {
+        urlPage = urlDemo + data["nextPage"];
+        judge = 1;
+        loadMore();
+        } else {
+        console.log("the last page");
       }
     });
-
-    function showLoading() {
-      setTimeout(getPost(), 2000);
-    }
-
-    function getPost() {
-      fetch(urlPage)
-        .then(function (response) {
-          return response.json();
-        })
-        .then(function (data) {
-          attractions = data["data"];
-          attractionLen = data["data"].length;
-          loadPicture();
-
-          if (data["nextPage"] != null) {
-            urlPage = urlDemo + data["nextPage"];
-            judge = 1;
-            loadMore();
-          } else {
-            console.log("the last page");
-          }
-        });
-    }
+  }
   } else {
     console.log("stop");
   }
@@ -170,8 +175,7 @@ function submit() {
   fetch("/api/categories")
     .then(function (response) {
       return response.json();
-    })
-    .then(function (data) {
+    }).then(function (data) {
       let categoriesLen = data["data"].length;
       for (let i = 0; i < categoriesLen; i++) {
         let searchItemNew = document.createElement("div");
@@ -187,339 +191,36 @@ function submit() {
         let itemName = "#" + data["data"][i];
         let itemSearchId = document.querySelector(itemName);
         itemSearchId.onclick = (e) => {
-          let search = document.querySelector(".search");
-          search.value = e.target.id;
-          searchContainer.style.display = "none";
+        let search = document.querySelector(".search");
+        search.value = e.target.id;
+        searchContainer.style.display = "none";
         };
       }
 
       let touchBody = document.querySelector("#body");
-      touchBody.addEventListener(
-        "click",
-        (e) => {
-          if (
-            e.target.className != "searchForm" &&
-            e.target.className != "searchItem" &&
-            e.target.className != "search"
-          ) {
-            searchContainer.style.display = "none";
-          }
-        },
-        false
-      );
-    });
+      touchBody.addEventListener("click",(e) => {
+      if(
+        e.target.className != "searchForm" &&
+        e.target.className != "searchItem" &&
+        e.target.className != "search"
+        ){
+        searchContainer.style.display = "none";
+        }},false);
+      });
 }
 
 //click icon -> clean imgBoxes -> return getDate()
 function icon() {
   let icon = document.querySelector(".icon");
-  icon.addEventListener(
-    "click",
-    (e) => {
-      let imgBoxes = document.querySelector(".imgBoxes");
-      while (imgBoxes.hasChildNodes()) {
-        imgBoxes.removeChild(imgBoxes.firstChild);
-      }
-      urlPage = urlKeywordDemo + search.value;
-      keyword = search.value;
-      getData();
-    },
-    false
-  );
+  icon.addEventListener("click",(e) => {
+    let imgBoxes = document.querySelector(".imgBoxes");
+    while (imgBoxes.hasChildNodes()) {
+    imgBoxes.removeChild(imgBoxes.firstChild);
+    }
+    urlPage = urlKeywordDemo + search.value;
+    keyword = search.value;
+    getData();
+  },false)
 }
 icon();
 
-// click  signup container alert
-let navSignup = document.querySelector(".nav-signup");
-navSignup.addEventListener(
-  "click",
-  (e) => {
-    let signupContainer = document.querySelector(".signupContainer");
-    signupContainer.style.display = "block";
-    let main = document.querySelector('#main');
-    main.style.opacity=.7;
-   
-  },
-  false
-);
-
-// control xmark
-let signupXmark = document.querySelector(".signupXmark");
-signupXmark.addEventListener(
-  "click",
-  (e) => {
-    let signupContainer = document.querySelector(".signupContainer");
-    signupContainer.style.display = "none";
-    main.style.opacity=1;
-  },
-  false
-);
-
-let signinXmark = document.querySelector(".signinXmark");
-signinXmark.addEventListener(
-  "click",
-  (e) => {
-    let signinContainer = document.querySelector(".signinContainer");
-    signinContainer.style.display = "none";
-    main.style.opacity=1;
-  },
-  false
-);
-
-//control signup and singin container page
-let signinToggle = document.querySelector(".signinToggle");
-signinToggle.addEventListener("click", (e) => {
-  let signupContainer = document.querySelector(".signupContainer");
-  signupContainer.style.display = "none";
-  let signinContainer = document.querySelector(".signinContainer");
-  signinContainer.style.display = "block";
-  main.style.opacity=.7;
-});
-
-let signupToggle = document.querySelector(".signupToggle");
-signupToggle.addEventListener("click", (e) => {
-  let signupContainer = document.querySelector(".signupContainer");
-  signupContainer.style.display = "block";
-  let signinContainer = document.querySelector(".signinContainer");
-  signinContainer.style.display = "none";
-});
-
-// signup (POST)
-let signupBtn = document.querySelector(".signupBtn");
-signupBtn.addEventListener("click", (e) => {
-  signupData();
-});
-
-function signupData() {
-  const name = document.getElementById("signup_name").value;
-  const email = document.getElementById("signup_email").value;
-  const password = document.getElementById("signup_password").value;
-
-  emailRule = /^\w+((-\w+)|(\.\w+))*\@[A-Za-z0-9]+((\.|-)[A-Za-z0-9]+)*\.[A-Za-z]+$/;
- 
-  //validate ok or not
-  if(email.search(emailRule)!= -1){
-    console.log('true');
-  }else{
-    let signupContainer = document.querySelector(".signupContainer");
-    signupContainer.style.height = "390px";
-        let signup_emailStatus = document.querySelector(".signup_emailStatus");
-        signup_emailStatus.innerHTML = "信箱格式輸入錯誤!";
-  }
-
-  const signupData = {
-    name: name,
-    email: email,
-    password: password,
-  };
-
-  fetch("/api/user", {
-    method: "POST",
-    credentials: "include",
-    body: JSON.stringify(signupData),
-    caches: "no-cache",
-    headers: new Headers({
-      "content-type": "application/json",
-    }),
-  })
-    .then(function (response) {
-      // if (response.status == 200){
-      // let signupStatus = document.querySelector(".signupStatus")
-      // signupStatus.innerHTML = '此信箱已註冊過，請重新註冊或登入'
-      // }
-      // console.log(response.status);
-      return response.json();
-    })
-    .then(function (data) {
-      // console.log(data);
-      // console.log(data[1]);
-      if (data['ok'] == true && email.search(emailRule)!= -1) {
-        let signupStatus = document.querySelector(".signupStatus");
-        signupStatus.innerHTML = "恭喜你！註冊成功！";
-      } else if (data['error'] == true) {
-        let signupContainer = document.querySelector(".signupContainer");
-        signupContainer.style.height = "375px";
-        let signupStatus = document.querySelector(".signupStatus");
-        signupStatus.innerHTML = "此信箱已註冊過，請重新輸入";
-      } else {
-        console.log("系統錯誤");
-      }
-    });
-}
-
-let signinBtn = document.querySelector(".signinBtn");
-signinBtn.addEventListener("click", (e) => {
-  signinData();
-  // getcookies();
-});
-
-//signin (PUT)
-function signinData() {
-  const email = document.getElementById("signin_email").value;
-  const password = document.getElementById("signin_password").value;
-
-  emailRule = /^\w+((-\w+)|(\.\w+))*\@[A-Za-z0-9]+((\.|-)[A-Za-z0-9]+)*\.[A-Za-z]+$/;
- 
-  //validate ok or not
-  if(email.search(emailRule)!= -1){
-    console.log('true');
-  }else{
-    let signinContainer = document.querySelector(".signinContainer");
-    signinContainer.style.height = "345px";
-        let signin_emailStatus = document.querySelector(".signin_emailStatus");
-        signin_emailStatus.innerHTML = "信箱格式輸入錯誤!";
-  }
-
-  const signinData = {
-    email: email,
-    password: password,
-  };
-
-  fetch("/api/user/auth", {
-    method: "PUT",
-    credentials: "include",
-    body: JSON.stringify(signinData),
-    caches: "no-cache",
-    headers: new Headers({
-      "content-type": "application/json",
-    }),
-  })
-    .then(function (response) {
-      // console.log(response);
-      return response.json();
-    })
-    .then(function (data) {
-      if (data['ok'] == true) {
-        // console.log('123');
-        // let navSignup = document.querySelector(".nav-signup")
-        // navSignup.innerHTML = '已登入'
-        let signinContainer = document.querySelector(".signinContainer");
-        signinContainer.style.display = "none";
-        main.style.opacity=1;
-        getcookie();
-      } else if (data["error"] == true) {
-        // signinContainer.style.display= 'none';
-        // let signinContainerPlus = document.querySelector(".signinContainerPlus");
-        // signinContainerPlus.style.display= 'block';
-        let signinContainer = document.querySelector(".signinContainer");
-        signinContainer.style.height = "320px";
-        let signinStatus = document.querySelector(".signinStatus");
-        signinStatus.innerHTML = "帳號密碼輸入錯誤，請重新輸入";
-      }
-    });
-}
-
-// function xhrr(){
-//   let xhr = new XMLHttpRequest();
-// xhr.open('get',"/api/user/auth",true)
-// xhr.send(null)
-// xhr.onload = function(){
-//  if (xhr.status === 200){
-//   let navsignup = document.querySelector(".nav-signup");
-//   navsignup.style.display = "none";
-//   let navsignout = document.querySelector(".nav-signout");
-//   navsignout.style.display = "block";
-
-//   console.log(xhr);
-//   console.log(xhr.response);
-//  }
-// }
-// }
-
-// console.log(xhr);
-
-//signin (GET)
-
-
-window.addEventListener('load',function(){
-  console.log('抓到你刷新頁面了嗎，讓我們檢查看看 token');
-  getcookie();
-})
-
-// getcookie();
-function getcookie() {
-  fetch("/api/user/auth", {
-    method: "GET",
-  })
-    .then(function (response) {
-      return response.json();
-    })
-    .then(function (data) {
-      console.log("取得token", data);
-      if (data["data"] != null) {
-        console.log("登入中狀態");
-        let navsignup = document.querySelector(".nav-signup");
-        navsignup.style.display = "none";
-        let navsignout = document.querySelector(".nav-signout");
-        navsignout.style.display = "block"; 
-      } 
-    });
-}
-
-// signoutData();
-function signoutData() {}
-  let navsignout = document.querySelector(".nav-signout");
-  navsignout.addEventListener("click", (e) => {
-    fetch("/api/user/auth", {
-      method: "DELETE",
-    })
-      .then(function (response) {
-        // console.log(response);
-        return response.json();
-      })
-
-      .then(function (data) {
-        console.log(data);
-        if (data['ok'] == true ) {
-          console.log("成功清除");
-          let navsignup = document.querySelector(".nav-signup");
-          navsignup.style.display = "block";
-          let navsignout = document.querySelector(".nav-signout");
-          navsignout.style.display = "none";
-        } else {
-          console.log("有問題");
-        }
-      });
-  });
-
-
-  //week-5
-  
-  //點擊預定行程頁面，
-let nav_booking = document.querySelector(".nav-booking")
-nav_booking.addEventListener('click',(e)=>{
-  console.log('確認登入狀態');
-  getcookie_navbooking();
-},false)
-
-function getcookie_navbooking(){
-  fetch("/api/user/auth", {
-    method: "GET",
-  })
-    .then(function (response) {
-      return response.json();
-    })
-    .then(function (data) {
-      console.log("取得token", data);
-      if (data["data"] != null) {
-        console.log("登入中狀態");
-        window.location.assign("http://172.20.10.2:3000/booking")
-      }
-      else{
-        //搜尋不到登入狀態 讓註冊登入視窗彈出
-        console.log('未登入狀態');
-        // window.location.replace(window.location.href);
-        let signupContainer = document.querySelector(".signupContainer");
-        signupContainer.style.display = "block";
-        let main = document.querySelector('#main');
-        main.style.opacity=.7;
-      } 
-    });
-} 
-
-//點擊 登出系統 跳回首頁
-// let navsignout = document.querySelector(".nav-signout"); 
-navsignout.addEventListener('click',(e)=>{
-  console.log('回首頁');
-  window.location.assign("http://172.20.10.2:3000/") 
-})
