+++ taipei-day-trip/static/css/base.css
@@ -6,6 +6,9 @@
   font-style: normal;
   text-decoration: none;
 }
+* ol, * ul, * li, * menu {
+  list-style: none;
+}
 
 body nav {
   width: 100%;
