+++ taipei-day-trip/static/css/base.scss
@@ -4,8 +4,12 @@
     padding:0;                      
     font-family: 'Noto Sans TC';    
     font-style: normal;      
-    text-decoration: none;       
+    text-decoration: none;  
+    ol, ul, li, menu{
+    list-style: none;
+}    
 }
+
 
 @mixin rwdFrist{
     @media(max-width:1250px){
@@ -33,6 +37,7 @@ body{
     nav{
         width:100%;
         height:54px;
+        //navbar 固定在頂部
         position: fixed;
         top:0;
         z-index: 2;
