+++ taipei-day-trip/static/css/index.css
@@ -6,6 +6,9 @@
   font-style: normal;
   text-decoration: none;
 }
+* ol, * ul, * li, * menu {
+  list-style: none;
+}
 
 body #main .pageOne {
   width: 100%;
