+++ taipei-day-trip/templates/attraction.html
@@ -62,7 +62,7 @@
 {% endblock %}
 
 {% block Javascript %}
-<script src="../static/attraction.js"></script>
-<script src="../static/booking.js"></script>
-    {% endblock %}
+<script type="module" src="../static/attraction.js"></script>
+<script type="module" src="../static/booking.js"></script>
+{% endblock %}
