+++ taipei-day-trip/templates/booking.html
@@ -5,7 +5,7 @@
 {% endblock %}
 
 {% block main %}
-<main>
+<main id="main">
   <section class="main">
     <img src="../static/img/icon_delete.png" alt="" class="icon_delete">
     <div class="page_top">
@@ -64,5 +64,5 @@
 {% endblock %}
 
 {% block Javascript %}
-<script src="../static/booking.js"></script>
+<script type="module" src="../static/booking.js"></script>
 {% endblock %}
