+++ taipei-day-trip/templates/index.html
@@ -39,5 +39,7 @@
 {% endblock %} 
 
 {% block Javascript %}
-<script src="../static/all.js"></script>
+<!-- <script src="../static/all.js"></script> -->
+<script type="module" src="../static/all.js"> </script>
+
 {% endblock %}
